feat: 하단 배너 광고 공통 위젯 분리 및 프로젝트 탭 적용
- 하단 고정 적응형 배너 광고 로직을 `lib/widgets/common_banner_ad.dart` 공통 위젯으로 분리
- `CommonBannerAdWidget` 내부에 `SafeArea` 및 `SizedBox`를 적용하여 레이아웃 안정성 확보
- `ProjectDetailScreen`의 기존 광고 로직을 공통 위젯으로 교체
- `ProjectsRoot` (프로젝트 탭) 하단에 공통 광고 위젯 적용
- 구글 공식 테스트 광고 ID 적용 (Android/iOS)
